% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{compute_predictions}
\alias{compute_predictions}
\title{compute_predictions}
\usage{
compute_predictions(model, new_data, level = 0.95)
}
\arguments{
\item{model}{A create_lm model object.}

\item{new_data}{The values of the predictor variable to make predictions for.}

\item{level}{A number between 0 and 1, the level of confidence.}
}
\value{
A tibble containing the predictions as well as response mean and prediction intervals.
}
\description{
Generates predictions for the response given a model and a vector of values of the predictor.
}
\author{
Tripp Bishop
}
