% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{create_lm}
\alias{create_lm}
\title{create_lm}
\usage{
create_lm(X, Y, level = 0.95)
}
\arguments{
\item{X}{A vector of the predictor variable.}

\item{Y}{A vector of the response variable.}

\item{level}{A number between 0 and 1, the level of confidence.}
}
\value{
A list containing parameters, confidence intervals of the parameters, residuals, and the coefficient of determination.
}
\description{
Computes the coefficients and residuals for a simple linear regression.
}
\author{
Tripp Bishop
}
