% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(model)
}
\arguments{
\item{model}{A create_lm model object.}
}
\value{
A ggplot2 object containing a plot of the model fit.
}
\description{
Generates a plot of model data and the best fit regression line determined by the model.
}
\author{
Tripp Bishop
}
