% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{plot_predictions}
\alias{plot_predictions}
\title{plot_predictions}
\usage{
plot_predictions(predictions, se = FALSE)
}
\arguments{
\item{predictions}{A tibble containing model prediction data.}

\item{se}{The mean and prediction intervals. Default is FALSE.}
}
\value{
A ggplot2 object containing a plot of the model predictions.
}
\description{
Generates a plot of model predictions and optionally confidence and prediction intervals.
}
\author{
Tripp Bishop
}
