% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(model)
}
\arguments{
\item{model}{A create_lm model object.}
}
\value{
A ggplot2 object containing a plot of the residual values of the create_lm model.
}
\description{
Generates a scatter plot of the residuals generated by the specified create_lm model object.
}
\author{
Tripp Bishop
}
