% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{poly_plot_fit}
\alias{poly_plot_fit}
\title{poly_plot_fit}
\usage{
poly_plot_fit(model)
}
\arguments{
\item{model}{A polynomial_lm model object.}
}
\value{
A ggplot2 object containing a plot of the model fit.
}
\description{
Generates a plot of model data and the best fit regression curve determined by the model.
}
\author{
Tripp Bishop
}
