% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{poly_plot_residuals}
\alias{poly_plot_residuals}
\title{poly_plot_residuals}
\usage{
poly_plot_residuals(model)
}
\arguments{
\item{model}{A polynomial_lm model object.}
}
\value{
A ggplot2 object containing a plot of the model fit.
}
\description{
Generates a scatter plot of the residuals generated by the specified create_lm model object.
}
\author{
Tripp Bishop
}
