% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{polynomial_lm}
\alias{polynomial_lm}
\title{polynomial_lm}
\usage{
polynomial_lm(X, Y, order = 2)
}
\arguments{
\item{X}{A vector of the predictor variable.}

\item{Y}{A vector of the response variable.}

\item{order}{An integer greater than or equal to 2. The order of polynomial to fit to the data.}
}
\value{
A list containing parameters, residuals, and the coefficient of determination.
}
\description{
Computes the coefficients and residuals for a simple polynomial linear regression.
}
\author{
Tripp Bishop
}
